requirements.txt
pandas==2.1.0
numpy==1.26.0
streamlit==1.22.0
streamlit-option-menu==0.3.6
tensorflow==2.15.0
pillow==10.1.0
scikit-learn==1.3.2
altair==4.2.0
protobuf==3.20.3